% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_model.R
\name{calc_mini_table_1}
\alias{calc_mini_table_1}
\title{calc_mini_table_1}
\usage{
calc_mini_table_1(prod_temp = 4, prod_days = 3, initial_conc = 0.2)
}
\description{
This function calculates the values that are seen in the
Lagring i produksjonsbedrift av laks (H14) mini table in the
Excel sheet.
}
\details{
It returns values that are equivalent to row 1 and row 2 of
the results table (A54) BUT without the cumulative sum.
}
