% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_model.R
\name{calc_mini_table_2}
\alias{calc_mini_table_2}
\title{calc_mini_table_2}
\usage{
calc_mini_table_2(store_temp = 4, store_days = 3, initial_conc = 0.2)
}
\description{
This function calculates the values that are seen in the
Innkjøp i butikk (H21) mini table in the  Excel sheet.
}
\details{
It returns values that are equivalent to row 3 of
the results table (A54) BUT without the cumulative sum.
}
