% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_model.R
\name{calc_mini_table_3_4}
\alias{calc_mini_table_3_4}
\title{calc_mini_table_3_4}
\usage{
calc_mini_table_3_4(
  prod_temp = 4,
  home_temp = 10,
  home_hours = 3,
  salmon_temp = 10,
  salmon_hours = 3,
  initial_conc = 0.2
)
}
\description{
This function calculates the values that are seen in the
Transport hjem (H29) and lagring av laks hjemme før laging av sushi
(H35) mini table in the  Excel sheet.
}
\details{
It returns values that are equivalent to row 4 and row 5 of
the results table (A54) BUT without the cumulative sum.

We do the two mini tables together because the reference growth for
the next step depends on the one in the previou step.
}
