% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_model.R
\name{calc_mini_table_5_6}
\alias{calc_mini_table_5_6}
\title{calc_mini_table_5_6}
\usage{
calc_mini_table_5_6(
  sushi_temp = 4,
  sushi_hours = 12,
  period_temp = 22,
  period_hours = 6,
  initial_conc = 0.2
)
}
\description{
This function calculates the values that are seen in the
Lagring av sushi hjemme, før temperering (H42) and
Tempereringsperiode (H48) mini table in the
Excel sheet.
}
\details{
It returns values that are equivalent to row 1 and row 2 of
the results table (A54) BUT without the cumulative sum.
}
